Add tests to relevant packages
