+++ src/test/java/no/unit/nva/model/instancetypes/book/BookAnthologyTest.java
@@ -1,8 +1,8 @@
-package no.unit.nva.model.instancetypes;
+package no.unit.nva.model.instancetypes.book;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import no.unit.nva.model.exceptions.InvalidPageRangeException;
-import no.unit.nva.model.instancetypes.book.BookMonograph;
+import no.unit.nva.model.instancetypes.InstanceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -10,31 +10,40 @@ import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class BookMonographTest extends InstanceTest {
+class BookAnthologyTest extends InstanceTest {
 
-    public static final String BOOK_MONOGRAPH = "BookMonograph";
+    public static final String BOOK_ANTHOLOGY = "BookAnthology";
 
-    @DisplayName("BookMonograph exists")
+    @DisplayName("BookAnthology exists")
     @Test
-    void bookMonographExists() {
-        new BookMonograph();
+    void bookAnthologyExists() {
+        new BookAnthology();
     }
 
-    @DisplayName("BookMonograph: ObjectMapper correctly deserializes object")
-    @ParameterizedTest(name = "BookMonograph deserializes begin {0}, end {1}, pages {2}, illustrated {3}, "
+    @DisplayName("BookAnthology: ObjectMapper correctly deserializes object")
+    @ParameterizedTest(name = "BookAnthology deserialized with begin {0}, end {1}, pages {2}, illustrated {3}, "
             + "peerReviewed {4}")
     @CsvSource({
             "i,xxviii,398,true,true",
             ",,231,false,true",
             ",,123,true,false"
     })
-    void objectMapperReturnsBookMonographWhenInputIsValid(String begin,
+    void objectMapperReturnsBookAnthologyWhenInputIsValid(String begin,
                                                           String end,
                                                           String pages,
                                                           boolean illustrated,
                                                           boolean peerReviewed) throws JsonProcessingException,
             InvalidPageRangeException {
-        BookMonograph expected = generateBookMonograph(
+
+        String json = generateMonographJsonString(
+                BOOK_ANTHOLOGY,
+                begin,
+                end,
+                pages,
+                illustrated,
+                peerReviewed);
+
+        BookAnthology expected = generateBookAnthology(
                 begin,
                 end,
                 pages,
@@ -42,20 +51,12 @@ class BookMonographTest extends InstanceTest {
                 peerReviewed
         );
 
-        String json = generateMonographJsonString(
-                BOOK_MONOGRAPH,
-                begin,
-                end,
-                pages,
-                illustrated,
-                peerReviewed
-        );
-        BookMonograph actual = objectMapper.readValue(json, BookMonograph.class);
+        BookAnthology actual = objectMapper.readValue(json, BookAnthology.class);
         assertEquals(expected, actual);
     }
 
-    @DisplayName("BookMonograph: ObjectMapper serializes valid input correctly")
-    @ParameterizedTest(name = "BookMonograph serializes begin {0}, end {1}, pages {2}, illustrated {3}, "
+    @DisplayName("BookAnthology: ObjectMapper serializes valid input correctly")
+    @ParameterizedTest(name = "BookAnthology serialized with begin {0}, end {1}, pages {2}, illustrated {3}, "
             + "peerReviewed {4}")
     @CsvSource({
             "i,xxviii,398,true,true",
@@ -66,28 +67,34 @@ class BookMonographTest extends InstanceTest {
                                                          String end,
                                                          String pages,
                                                          boolean illustrated,
-                                                         boolean peerReviewed) throws
-            InvalidPageRangeException, JsonProcessingException {
+                                                         boolean peerReviewed) throws JsonProcessingException,
+            InvalidPageRangeException {
 
-        BookMonograph bookMonograph = generateBookMonograph(begin,
+        BookAnthology bookAnthology = generateBookAnthology(
+                begin,
                 end,
                 pages,
                 illustrated,
                 peerReviewed
         );
-        String json = objectMapper.writeValueAsString(bookMonograph);
-        String expected = generateMonographJsonString(BOOK_MONOGRAPH,
-                begin, end, pages, illustrated, peerReviewed);
+        String json = objectMapper.writeValueAsString(bookAnthology);
+        String expected = generateMonographJsonString(
+                BOOK_ANTHOLOGY,
+                begin,
+                end,
+                pages,
+                illustrated,
+                peerReviewed);
         assertEquals(expected, json);
     }
 
-    private BookMonograph generateBookMonograph(String introductionBegin,
+    private BookAnthology generateBookAnthology(String introductionBegin,
                                                 String introductionEnd,
                                                 String pages,
                                                 boolean illustrated,
                                                 boolean peerReviewed) throws InvalidPageRangeException {
 
-        return new BookMonograph.Builder()
+        return new BookAnthology.Builder()
                 .withPages(generateMonographPages(introductionBegin, introductionEnd, pages, illustrated))
                 .withPeerReviewed(peerReviewed)
                 .build();
