+++ src/test/java/no/unit/nva/model/instancetypes/book/BookMonographTest.java
@@ -1,8 +1,8 @@
-package no.unit.nva.model.instancetypes;
+package no.unit.nva.model.instancetypes.book;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import no.unit.nva.model.exceptions.InvalidPageRangeException;
-import no.unit.nva.model.instancetypes.book.BookAnthology;
+import no.unit.nva.model.instancetypes.InstanceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -10,40 +10,31 @@ import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class BookAnthologyTest extends InstanceTest {
+class BookMonographTest extends InstanceTest {
 
-    public static final String BOOK_ANTHOLOGY = "BookAnthology";
+    public static final String BOOK_MONOGRAPH = "BookMonograph";
 
-    @DisplayName("BookAnthology exists")
+    @DisplayName("BookMonograph exists")
     @Test
-    void bookAnthologyExists() {
-        new BookAnthology();
+    void bookMonographExists() {
+        new BookMonograph();
     }
 
-    @DisplayName("BookAnthology: ObjectMapper correctly deserializes object")
-    @ParameterizedTest(name = "BookAnthology deserialized with begin {0}, end {1}, pages {2}, illustrated {3}, "
+    @DisplayName("BookMonograph: ObjectMapper correctly deserializes object")
+    @ParameterizedTest(name = "BookMonograph deserializes begin {0}, end {1}, pages {2}, illustrated {3}, "
             + "peerReviewed {4}")
     @CsvSource({
             "i,xxviii,398,true,true",
             ",,231,false,true",
             ",,123,true,false"
     })
-    void objectMapperReturnsBookAnthologyWhenInputIsValid(String begin,
+    void objectMapperReturnsBookMonographWhenInputIsValid(String begin,
                                                           String end,
                                                           String pages,
                                                           boolean illustrated,
                                                           boolean peerReviewed) throws JsonProcessingException,
             InvalidPageRangeException {
-
-        String json = generateMonographJsonString(
-                BOOK_ANTHOLOGY,
-                begin,
-                end,
-                pages,
-                illustrated,
-                peerReviewed);
-
-        BookAnthology expected = generateBookAnthology(
+        BookMonograph expected = generateBookMonograph(
                 begin,
                 end,
                 pages,
@@ -51,12 +42,20 @@ class BookAnthologyTest extends InstanceTest {
                 peerReviewed
         );
 
-        BookAnthology actual = objectMapper.readValue(json, BookAnthology.class);
+        String json = generateMonographJsonString(
+                BOOK_MONOGRAPH,
+                begin,
+                end,
+                pages,
+                illustrated,
+                peerReviewed
+        );
+        BookMonograph actual = objectMapper.readValue(json, BookMonograph.class);
         assertEquals(expected, actual);
     }
 
-    @DisplayName("BookAnthology: ObjectMapper serializes valid input correctly")
-    @ParameterizedTest(name = "BookAnthology serialized with begin {0}, end {1}, pages {2}, illustrated {3}, "
+    @DisplayName("BookMonograph: ObjectMapper serializes valid input correctly")
+    @ParameterizedTest(name = "BookMonograph serializes begin {0}, end {1}, pages {2}, illustrated {3}, "
             + "peerReviewed {4}")
     @CsvSource({
             "i,xxviii,398,true,true",
@@ -67,34 +66,28 @@ class BookAnthologyTest extends InstanceTest {
                                                          String end,
                                                          String pages,
                                                          boolean illustrated,
-                                                         boolean peerReviewed) throws JsonProcessingException,
-            InvalidPageRangeException {
+                                                         boolean peerReviewed) throws
+            InvalidPageRangeException, JsonProcessingException {
 
-        BookAnthology bookAnthology = generateBookAnthology(
-                begin,
+        BookMonograph bookMonograph = generateBookMonograph(begin,
                 end,
                 pages,
                 illustrated,
                 peerReviewed
         );
-        String json = objectMapper.writeValueAsString(bookAnthology);
-        String expected = generateMonographJsonString(
-                BOOK_ANTHOLOGY,
-                begin,
-                end,
-                pages,
-                illustrated,
-                peerReviewed);
+        String json = objectMapper.writeValueAsString(bookMonograph);
+        String expected = generateMonographJsonString(BOOK_MONOGRAPH,
+                begin, end, pages, illustrated, peerReviewed);
         assertEquals(expected, json);
     }
 
-    private BookAnthology generateBookAnthology(String introductionBegin,
+    private BookMonograph generateBookMonograph(String introductionBegin,
                                                 String introductionEnd,
                                                 String pages,
                                                 boolean illustrated,
                                                 boolean peerReviewed) throws InvalidPageRangeException {
 
-        return new BookAnthology.Builder()
+        return new BookMonograph.Builder()
                 .withPages(generateMonographPages(introductionBegin, introductionEnd, pages, illustrated))
                 .withPeerReviewed(peerReviewed)
                 .build();
