+++ src/test/java/no/unit/nva/model/instancetypes/report/ReportWorkingPaperTest.java
@@ -1,9 +1,9 @@
-package no.unit.nva.model.instancetypes;
+package no.unit.nva.model.instancetypes.report;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import no.unit.nva.model.exceptions.InvalidPageRangeException;
-import no.unit.nva.model.instancetypes.report.ReportPolicy;
+import no.unit.nva.model.instancetypes.InstanceTest;
 import no.unit.nva.model.pages.MonographPages;
 import no.unit.nva.model.pages.Range;
 import nva.commons.utils.JsonUtils;
@@ -17,54 +17,57 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class ReportPolicyTest extends InstanceTest {
+class ReportWorkingPaperTest extends InstanceTest {
 
     public static final ObjectMapper objectMapper = JsonUtils.objectMapper;
 
-    @DisplayName("Report policy can be created from JSON")
+    @DisplayName("Report working paper can be created from JSON")
     @Test
-    void reportPolicyReturnsObjectWhenJsonInputIsCorrectlySerialized() throws JsonProcessingException,
+    void reportWorkingPaperReturnsObjectWhenJsonInputIsCorrectlySerialized() throws JsonProcessingException,
             InvalidPageRangeException {
-        ReportPolicy expected = generateReportPolicy("42", "1", "3", false);
+        ReportWorkingPaper expected = generateReportWorkingPaper("42", "1", "3", false);
         String json = objectMapper.writeValueAsString(expected);
         assertTrue(json.contains("peerReviewed"));
-        ReportPolicy reportPolicy = objectMapper.readValue(json, ReportPolicy.class);
-        assertEquals(expected, reportPolicy);
+        ReportWorkingPaper reportWorkingPaper = objectMapper.readValue(json, ReportWorkingPaper.class);
+        assertEquals(expected, reportWorkingPaper);
     }
 
-    @DisplayName("Report policy cannot be peer reviewed")
+    @DisplayName("Report working paper cannot be peer reviewed")
     @Test
-    void reportPolicySetsPeerReviewedToFalseWhenPeerReviewIsTrue() throws JsonProcessingException,
+    void reportWorkingPaperSetsPeerReviewedToFalseWhenPeerReviewIsTrue() throws JsonProcessingException,
             InvalidPageRangeException {
-        String type = "ReportPolicy";
+        ObjectMapper objectMapper = new ObjectMapper();
+        String type = "ReportWorkingPaper";
         String pages = "42";
         String introductionBegin = "1";
         String introductionEnd = "3";
         boolean illustrated = false;
         boolean peerReviewed = true;
-        ReportPolicy expected = generateReportPolicy(pages, introductionBegin, introductionEnd, illustrated);
+        ReportWorkingPaper expected = generateReportWorkingPaper(pages, introductionBegin, introductionEnd,
+                illustrated);
 
         String json = generateMonographJsonString(type, introductionBegin, introductionEnd, pages,
                 illustrated, peerReviewed);
-        ReportPolicy reportPolicy = objectMapper.readValue(json, ReportPolicy.class);
-        assertEquals(expected, reportPolicy);
+        ReportWorkingPaper reportWorkingPaper = objectMapper.readValue(json, ReportWorkingPaper.class);
+        assertEquals(expected, reportWorkingPaper);
     }
 
-    @DisplayName("ReportPolicy: Attempting to set peer reviewed to true results in Unexpected exception")
+    @DisplayName("ReportWorkingPaper: Attempting to set peer reviewed to true results in Unexpected exception")
     @Test
-    void reportThrowsUnexpectedExceptionWhenPeerReviewedIsTrue() {
+    void reportWorkingPaperThrowsUnexpectedExceptionWhenPeerReviewedIsTrue() {
         Executable executable = () -> {
-            ReportPolicy reportPolicy = new ReportPolicy();
-            reportPolicy.setPeerReviewed(true);
+            ReportWorkingPaper reportWorkingPaper = new ReportWorkingPaper();
+            reportWorkingPaper.setPeerReviewed(true);
         };
         UnexpectedException exception = assertThrows(UnexpectedException.class, executable);
-        String expected = String.format(ReportPolicy.PEER_REVIEWED_ERROR_TEMPLATE,
-                ReportPolicy.class.getSimpleName());
+        String expected = String.format(ReportWorkingPaper.PEER_REVIEWED_ERROR_TEMPLATE,
+                ReportWorkingPaper.class.getSimpleName());
         assertEquals(expected, exception.getMessage());
     }
 
-    private ReportPolicy generateReportPolicy(String pages, String introductionBegin, String introductionEnd,
-                                              boolean illustrated) throws InvalidPageRangeException {
+    private ReportWorkingPaper generateReportWorkingPaper(String pages, String introductionBegin,
+        String introductionEnd, boolean illustrated) throws InvalidPageRangeException {
+
         Range introductionRange = new Range.Builder()
                 .withBegin(introductionBegin)
                 .withEnd(introductionEnd)
@@ -76,7 +79,7 @@ class ReportPolicyTest extends InstanceTest {
                 .withIllustrated(illustrated)
                 .build();
 
-        return new ReportPolicy.Builder()
+        return new ReportWorkingPaper.Builder()
                 .withPages(monographPages)
                 .build();
     }
